fix #170 论文：增加毕业论文归档信息

+++ degree/src/main/scala/org/openurp/degree/thesis/model/BlindPeerReview.scala
@@ -28,6 +28,6 @@ class BlindPeerReview extends LongId, Updated, Remark {
   var writer: Writer = _
 
   /** 分数 */
-  var score: Option[Float] = None
+  var score: Option[Int] = None
 
 }
